Enable cast_unittests And fixed associated errors
Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=228002

Review URL: https://codereview.chromium.org/26486005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@228515 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/cast/audio_receiver/audio_decoder_unittest.cc
@@ -44,10 +44,10 @@ TEST_F(AudioDecoderTest, Pcm16MonoNoResampleOnePacket) {
   std::vector<int16> payload(640, 0x1234);
 
   uint8* payload_data = reinterpret_cast<uint8*>(&payload[0]);
-  int payload_size = payload.size() * sizeof(int16);
+  size_t payload_size = payload.size() * sizeof(int16);
 
-  audio_decoder_->IncomingParsedRtpPacket(payload_data, payload_size,
-                                          rtp_header);
+  audio_decoder_->IncomingParsedRtpPacket(payload_data,
+      static_cast<int>(payload_size), rtp_header);
 
   int number_of_10ms_blocks = 4;
   int desired_frequency = 16000;
@@ -91,10 +91,10 @@ TEST_F(AudioDecoderTest, Pcm16StereoNoResampleTwoPackets) {
   std::vector<int16> payload(640, 0x1234);
 
   uint8* payload_data = reinterpret_cast<uint8*>(&payload[0]);
-  int payload_size = payload.size() * sizeof(int16);
+  size_t payload_size = payload.size() * sizeof(int16);
 
-  audio_decoder_->IncomingParsedRtpPacket(payload_data, payload_size,
-                                          rtp_header);
+  audio_decoder_->IncomingParsedRtpPacket(payload_data,
+      static_cast<int>(payload_size), rtp_header);
 
 
   int number_of_10ms_blocks = 2;
@@ -156,10 +156,10 @@ TEST_F(AudioDecoderTest, Pcm16Resample) {
   std::vector<int16> payload(640, 0x1234);
 
   uint8* payload_data = reinterpret_cast<uint8*>(&payload[0]);
-  int payload_size = payload.size() * sizeof(int16);
+  size_t payload_size = payload.size() * sizeof(int16);
 
-  audio_decoder_->IncomingParsedRtpPacket(payload_data, payload_size,
-                                          rtp_header);
+  audio_decoder_->IncomingParsedRtpPacket(payload_data,
+      static_cast<int>(payload_size), rtp_header);
 
   int number_of_10ms_blocks = 2;
   int desired_frequency = 48000;
